tuesday fixes

+++ application/modules/clients/controllers/clients.php
@@ -7,12 +7,12 @@ class Clients extends CI_Controller{
 		$this->load->model("clients/clients_model");
 		$this->load->model("user/user_model");
 	}
-	function index($msr_id,$offset = 0) {
+	function index($msr_id =0,$offset = 0) {
 		$this->display($msr_id,$offset);
 	}
 	function display($msr_id,$offset = 0) {
 		$this->load->library('pagination');
-        if($this->users->is_admin() || $this->users->is_msr() || $this->users->is_accountant()) {
+        if($this->users->is_admin() || $this->users->is_msr() || $this->users->is_accountant() || $this->users->is_hrd()) {
             if($this->users->is_admin()) {
 				$config['base_url'] = base_url().'clients/display';
 				$config['total_rows'] = $this->clients_model->count_all();
@@ -20,7 +20,7 @@ class Clients extends CI_Controller{
 				$this->pagination->initialize($config); 
 				$data['client'] = $this->clients_model->get_all($offset,$config['per_page']);
 				$this->template->load('template','clients/clients_list',$data);
-			} else if($this->users->is_accountant() || $this->users->is_msr()){
+			} else if($this->users->is_accountant() || $this->users->is_msr() || $this->users->is_hrd()){
 				$config['base_url'] = base_url().'clients/display';
 				$config['total_rows'] = $this->client_model->get_clients($msr_id,true);
 				$config['per_page'] = 15;
@@ -34,7 +34,7 @@ class Clients extends CI_Controller{
         }
 	}
 	function create() {
-        if($this->users->is_admin() || $this->users->is_msr()) {
+        if($this->users->is_admin() || $this->users->is_msr() || $this->users->is_hrd()) {
     		if(!empty($_POST)){
     			if($this->user_model->get_single($_POST['username'],true,'username') == 'Invalid user') { 
 					$id = $this->user_model->create($_POST);
@@ -64,17 +64,19 @@ class Clients extends CI_Controller{
         }
 	}
 	function edit($id) {
-		if(!empty($_POST)) {
-			$this->client_model->update($id,$_POST);
-			$this->session->set_flashdata('error','client updated!');	
-			redirect('client');
-		} else {
-			$data['client'] = $this->client_model->get_single($id);
-			if(!$data['client']) {
-				$this->session->set_flashdata('error','not a valid client!');	
+		if($this->users->is_admin() || $this->users->is_msr() || $this->users->is_hrd()) {
+			if(!empty($_POST)) {
+				$this->client_model->update($id,$_POST);
+				$this->session->set_flashdata('error','client updated!');	
 				redirect('client');
+			} else {
+				$data['client'] = $this->client_model->get_single($id);
+				if(!$data['client']) {
+					$this->session->set_flashdata('error','not a valid client!');	
+					redirect('client');
+				}
+				$this->template->load('template','clients/client_edit',$data);
 			}
-			$this->template->load('template','clients/client_edit',$data);
 		}
 	}
 	function view($id) {
@@ -86,11 +88,13 @@ class Clients extends CI_Controller{
 		$this->template->load('template','clients/client_view',$data);
 	}
 	function delete($id) {
-		if(!empty($id)) {
-			if($this->client_model->delete($id)) {
-				$this->session->set_flashdata('error','client removed!');	
-			} else {
-				$this->session->set_flashdata('error','client not removed!');	
+		if($this->users->is_admin() || $this->users->is_msr() || $this->users->is_hrd()) {
+			if(!empty($id)) {
+				if($this->client_model->delete($id)) {
+					$this->session->set_flashdata('error','client removed!');	
+				} else {
+					$this->session->set_flashdata('error','client not removed!');	
+				}
 			}
 		}
 		redirect('client');
